Removed Cmake, prototxt files, added new fp16
