--- apps/gender_age_lbp/build.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-
-mkdir -p build
-cd build
-cmake ..
-make
