+++ apps/gender_age_lbp/catstat/Gender/categories.txt
@@ -1,0 +1,3 @@
+classes
+Male
+Female
