+++ apps/gender_age_lbp/cpp/fp16.h
@@ -1,9 +1,9 @@
+// Copied from Numpy
 
-#ifndef INCLUDE_FP16_H
 
-unsigned half2float(unsigned short h);
+static unsigned half2float(unsigned short h);
 unsigned short float2half(unsigned f);
+void floattofp16(unsigned char *dst, float *src, unsigned nelem);
 void fp16tofloat(float *dst, unsigned char *src, unsigned nelem);
 
 
-#endif
