--- apps/gender_age_lbp/networks/Age/categories.txt
@@ -1,9 +1,0 @@
-class
-0-2
-4-6
-8-12
-15-20
-25-32
-38-43
-48-53
-60-100
