--- apps/gender_age_lbp/networks/Gender/categories.txt
@@ -1,3 +1,0 @@
-classes
-Male
-Female
