--- apps/gender_age_lbp/networks/Gender/stat.txt
@@ -1,2 +1,0 @@
-0.3075542657268599 0.34419182107610374 0.45057195194485578
-0.00392156862745098 0.00392156862745098 0.00392156862745098
